iot_pipeline/requirements.txt
# Core Kafka and Database
kafka-python==2.0.2
psycopg2-binary==2.9.7
sqlalchemy==2.0.20

# Data Processing
pandas==2.0.3
numpy==1.24.3

# Machine Learning
scikit-learn==1.3.0

# Visualization
streamlit==1.26.0
plotly==5.16.1

# Utilities
python-dateutil==2.8.2
